requirements_1.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.5.7
dataclasses==0.6
dj-database-url==2.0.0
Django==4.1.13
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.4.0
exceptiongroup==1.1.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
packaging==23.1
psycopg2==2.9.6
PyJWT==2.8.0
pymongo==3.12.1
pytz==2023.3
sniffio==1.3.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
